Modify validation/options handling
