+++ tests/addons/test_options_schema.py
@@ -3,29 +3,29 @@
 import pytest
 import voluptuous as vol
 
-from supervisor.addons.validate import validate_options
+from supervisor.addons.options import AddonOptions
 
 
 def test_simple_schema(coresys):
     """Test with simple schema."""
-    assert validate_options(
+    assert AddonOptions(
         coresys,
         {"name": "str", "password": "password", "fires": "bool", "alias": "str?"},
     )({"name": "Pascal", "password": "1234", "fires": True, "alias": "test"})
 
-    assert validate_options(
+    assert AddonOptions(
         coresys,
         {"name": "str", "password": "password", "fires": "bool", "alias": "str?"},
     )({"name": "Pascal", "password": "1234", "fires": True})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "fires": "bool", "alias": "str?"},
         )({"name": "Pascal", "password": "1234", "fires": "hah"})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "fires": "bool", "alias": "str?"},
         )({"name": "Pascal", "fires": True})
@@ -33,19 +33,19 @@ def test_simple_schema(coresys):
 
 def test_complex_schema_list(coresys):
     """Test with complex list schema."""
-    assert validate_options(
+    assert AddonOptions(
         coresys,
         {"name": "str", "password": "password", "extend": ["str"]},
     )({"name": "Pascal", "password": "1234", "extend": ["test", "blu"]})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "extend": ["str"]},
         )({"name": "Pascal", "password": "1234", "extend": ["test", 1]})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "extend": ["str"]},
         )({"name": "Pascal", "password": "1234", "extend": "test"})
@@ -53,19 +53,19 @@ def test_complex_schema_list(coresys):
 
 def test_complex_schema_dict(coresys):
     """Test with complex dict schema."""
-    assert validate_options(
+    assert AddonOptions(
         coresys,
         {"name": "str", "password": "password", "extend": {"test": "int"}},
     )({"name": "Pascal", "password": "1234", "extend": {"test": 1}})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "extend": {"test": "int"}},
         )({"name": "Pascal", "password": "1234", "extend": {"wrong": 1}})
 
     with pytest.raises(vol.error.Invalid):
-        validate_options(
+        AddonOptions(
             coresys,
             {"name": "str", "password": "password", "extend": ["str"]},
         )({"name": "Pascal", "password": "1234", "extend": "test"})
